added pop ups

+++ frontend/src/pages/encoder.js
@@ -85,6 +85,18 @@ function Encoder() {
     fetchNotes();
   }, []);
 
+
+
+  const send_data = () => {
+    const confirm_send = window.confirm("Send data for approval?")
+    if (confirm_send){
+      alert("Data sent for approval!")
+    }else {
+      alert("Data not sent for approval!")
+    }
+  }
+
+
   const fetchNotes = async () => {
     //fetch notes
     const res = await axios.get("http://localhost:3001/data");
@@ -151,8 +163,16 @@ function Encoder() {
   };
 
   const deleteNote = async (_id) => {
-    const res = await axios.delete(`http://localhost:3001/data/${_id}`);
-    fetchNotes();
+    const confirm = window.confirm(
+      "Are you sure you want to delete this record?"
+    );
+    if (confirm) {
+      const res = await axios.delete(`http://localhost:3001/data/${_id}`);
+      fetchNotes();
+    } else {
+      alert("Record not deleted!");
+    }
+
     //console.log(res)
   };
 
@@ -228,113 +248,117 @@ function Encoder() {
   const update = async (e) => {
     // Prevent default form submit
     e.preventDefault();
+    const confirm_update = window.confirm("Confirm record update?");
+  
+    if (confirm_update) {
+      // Construct updated record
+      const updatedRecord = {
+        Year: updateForm.Year,
+        Month: updateForm.Month,
+        "Water Intake (Purchased)": updateForm["Water Intake (Purchased)"],
+        "Water Use (Sweat) - Use": updateForm["Water Use (Sweat) - Use"],
+        "Water Loss": updateForm["Water Loss"],
+        "Wastewater Disposal (Outside)":
+          updateForm["Wastewater Disposal (Outside)"],
+        "Total Energy": updateForm["Total Energy"],
+        "Energy Pruchased (Electricity)":
+          updateForm["Energy Pruchased (Electricity)"],
+        "Energy (Diesel)": updateForm["Energy (Diesel)"],
+        "Total Waste": updateForm["Total Waste"],
+        "Total Hazardous Waste (Landfil)":
+          updateForm["Total Hazardous Waste (Landfil)"],
+        "Hazardous Waste (Stream # 1)":
+          updateForm["Hazardous Waste (Stream # 1)"],
+        "Hazardous Waste (Maintenace Oil) - (With Density 0.93)":
+          updateForm["Hazardous Waste (Maintenace Oil) - (With Density 0.93)"],
+        "Total Non-Hazardous ( Landfill + Recycling)":
+          updateForm["Total Non-Hazardous ( Landfill + Recycling)"],
+        "Non-Hazardous Waste (Scrap Recycling))":
+          updateForm["Non-Hazardous Waste (Scrap Recycling))"],
+        "Non-Hazardous Waste (Disposal)":
+          updateForm["Non-Hazardous Waste (Disposal)"],
+        "Total GHG (CO2)": updateForm["Total GHG (CO2)"],
+        "Total Consumption R22 Gas": updateForm["Total Consumption R22 Gas"],
+        "Cost of Water (Sweat) Purchase / Treatment / Water Waste Disposal":
+          updateForm[
+            "Cost of Water (Sweat) Purchase / Treatment / Water Waste Disposal"
+          ],
+        "Cost of Pruchased Electricity":
+          updateForm["Cost of Pruchased Electricity"],
+        "Cost of Diesel": updateForm["Cost of Diesel"],
+        "Cost of R22 Gas": updateForm["Cost of R22 Gas"],
+        "Cost of Hazardouse Waste Disposal":
+          updateForm["Cost of Hazardouse Waste Disposal"],
+        "Cost of Non-Hazardouse Waste Disposal":
+          updateForm["Cost of Non-Hazardouse Waste Disposal"],
+        "Total Production Volume (Pallets Issue)":
+          updateForm["Total Production Volume (Pallets Issue)"],
+        "Total Production Volume (Pallets Receive)":
+          updateForm["Total Production Volume (Pallets Receive)"],
+        "Total Area (Floor Covered Area)":
+          updateForm["Total Area (Floor Covered Area)"],
+        "Intensity - Water": updateForm["Intensity - Water"],
+        "Intensity - Energy": updateForm["Intensity - Energy"],
+        "Intensity - Energy (Electricity)":
+          updateForm["Intensity - Energy (Electricity)"],
+        "Intensity - Energy (Diesel)":
+          updateForm["Intensity - Energy (Diesel)"],
+        "Intensity - GHG (CO2) - Scope 1 and 2":
+          updateForm["Intensity - GHG (CO2) - Scope 1 and 2"],
+        "Intensity - Waste Material": updateForm["Intensity - Waste Material"],
+        "Total Cost": updateForm["Total Cost"],
+        Comment: updateForm.Comment,
+      };
 
-    // Construct updated record
-    const updatedRecord = {
-      Year: updateForm.Year,
-      Month: updateForm.Month,
-      "Water Intake (Purchased)": updateForm["Water Intake (Purchased)"],
-      "Water Use (Sweat) - Use": updateForm["Water Use (Sweat) - Use"],
-      "Water Loss": updateForm["Water Loss"],
-      "Wastewater Disposal (Outside)":
-        updateForm["Wastewater Disposal (Outside)"],
-      "Total Energy": updateForm["Total Energy"],
-      "Energy Pruchased (Electricity)":
-        updateForm["Energy Pruchased (Electricity)"],
-      "Energy (Diesel)": updateForm["Energy (Diesel)"],
-      "Total Waste": updateForm["Total Waste"],
-      "Total Hazardous Waste (Landfil)":
-        updateForm["Total Hazardous Waste (Landfil)"],
-      "Hazardous Waste (Stream # 1)":
-        updateForm["Hazardous Waste (Stream # 1)"],
-      "Hazardous Waste (Maintenace Oil) - (With Density 0.93)":
-        updateForm["Hazardous Waste (Maintenace Oil) - (With Density 0.93)"],
-      "Total Non-Hazardous ( Landfill + Recycling)":
-        updateForm["Total Non-Hazardous ( Landfill + Recycling)"],
-      "Non-Hazardous Waste (Scrap Recycling))":
-        updateForm["Non-Hazardous Waste (Scrap Recycling))"],
-      "Non-Hazardous Waste (Disposal)":
-        updateForm["Non-Hazardous Waste (Disposal)"],
-      "Total GHG (CO2)": updateForm["Total GHG (CO2)"],
-      "Total Consumption R22 Gas": updateForm["Total Consumption R22 Gas"],
-      "Cost of Water (Sweat) Purchase / Treatment / Water Waste Disposal":
-        updateForm[
-        "Cost of Water (Sweat) Purchase / Treatment / Water Waste Disposal"
-        ],
-      "Cost of Pruchased Electricity":
-        updateForm["Cost of Pruchased Electricity"],
-      "Cost of Diesel": updateForm["Cost of Diesel"],
-      "Cost of R22 Gas": updateForm["Cost of R22 Gas"],
-      "Cost of Hazardouse Waste Disposal":
-        updateForm["Cost of Hazardouse Waste Disposal"],
-      "Cost of Non-Hazardouse Waste Disposal":
-        updateForm["Cost of Non-Hazardouse Waste Disposal"],
-      "Total Production Volume (Pallets Issue)":
-        updateForm["Total Production Volume (Pallets Issue)"],
-      "Total Production Volume (Pallets Receive)":
-        updateForm["Total Production Volume (Pallets Receive)"],
-      "Total Area (Floor Covered Area)":
-        updateForm["Total Area (Floor Covered Area)"],
-      "Intensity - Water": updateForm["Intensity - Water"],
-      "Intensity - Energy": updateForm["Intensity - Energy"],
-      "Intensity - Energy (Electricity)":
-        updateForm["Intensity - Energy (Electricity)"],
-      "Intensity - Energy (Diesel)": updateForm["Intensity - Energy (Diesel)"],
-      "Intensity - GHG (CO2) - Scope 1 and 2":
-        updateForm["Intensity - GHG (CO2) - Scope 1 and 2"],
-      "Intensity - Waste Material": updateForm["Intensity - Waste Material"],
-      "Total Cost": updateForm["Total Cost"],
-      Comment: updateForm.Comment,
-    };
+      // Make PUT request to update
+      const res = await axios.put(
+        `http://localhost:3001/data/${updateForm._id}`,
+        updatedRecord
+      );
 
-    // Make PUT request to update
-    const res = await axios.put(
-      `http://localhost:3001/data/${updateForm._id}`,
-      updatedRecord
-    );
+      fetchNotes();
 
-    fetchNotes();
-
-    setUpdateForm({
-      _id: null,
-      Year: "",
-      Month: "",
-      "Water Intake (Purchased)": "",
-      "Water Use (Sweat) - Use": "",
-      "Water Loss": "",
-      "Wastewater Disposal (Outside)": "",
-      "Total Energy": "",
-      "Energy Pruchased (Electricity)": "",
-      "Energy (Diesel)": "",
-      "Total Waste": "",
-      "Total Hazardous Waste (Landfil)": "",
-      "Hazardous Waste (Stream # 1)": "",
-      "Hazardous Waste (Maintenace Oil)(With Density 0_93)": "",
-      "Total Non-Hazardous ( Landfill + Recycling)": "",
-      "Non-Hazardous Waste (Scrap Recycling))": "",
-      "Non-Hazardous Waste (Disposal)": "",
-      "Total GHG (CO2)": "",
-      "Total Consumption R22 Gas": "",
-      "Cost of Water (Sweat) Purchase_Treatment_Water Waste_Disposal": "",
-      "Cost of Pruchased Electricity": "",
-      "Cost of Diesel": "",
-      "Cost of R22 Gas": "",
-      "Cost of Hazardouse Waste Disposal": "",
-      "Cost of Non-Hazardouse Waste Disposal": "",
-      "Total Production Volume (Pallets Issue)": "",
-      "Total Production Volume (Pallets Receive)": "",
-      "Total Area (Floor Covered Area)": "",
-      "Intensity - Water": "",
-      "Intensity - Energy": "",
-      "Intensity - Energy (Electricity)": "",
-      "Intensity - Energy (Diesel)": "",
-      "Intensity - GHG (CO2) - Scope 1 and 2": "",
-      "Intensity - Waste Material": "",
-      "Total Cost": "",
-      Comment: "",
-    });
-
-    console.log(res.data);
+      setUpdateForm({
+        _id: null,
+        Year: "",
+        Month: "",
+        "Water Intake (Purchased)": "",
+        "Water Use (Sweat) - Use": "",
+        "Water Loss": "",
+        "Wastewater Disposal (Outside)": "",
+        "Total Energy": "",
+        "Energy Pruchased (Electricity)": "",
+        "Energy (Diesel)": "",
+        "Total Waste": "",
+        "Total Hazardous Waste (Landfil)": "",
+        "Hazardous Waste (Stream # 1)": "",
+        "Hazardous Waste (Maintenace Oil)(With Density 0_93)": "",
+        "Total Non-Hazardous ( Landfill + Recycling)": "",
+        "Non-Hazardous Waste (Scrap Recycling))": "",
+        "Non-Hazardous Waste (Disposal)": "",
+        "Total GHG (CO2)": "",
+        "Total Consumption R22 Gas": "",
+        "Cost of Water (Sweat) Purchase_Treatment_Water Waste_Disposal": "",
+        "Cost of Pruchased Electricity": "",
+        "Cost of Diesel": "",
+        "Cost of R22 Gas": "",
+        "Cost of Hazardouse Waste Disposal": "",
+        "Cost of Non-Hazardouse Waste Disposal": "",
+        "Total Production Volume (Pallets Issue)": "",
+        "Total Production Volume (Pallets Receive)": "",
+        "Total Area (Floor Covered Area)": "",
+        "Intensity - Water": "",
+        "Intensity - Energy": "",
+        "Intensity - Energy (Electricity)": "",
+        "Intensity - Energy (Diesel)": "",
+        "Intensity - GHG (CO2) - Scope 1 and 2": "",
+        "Intensity - Waste Material": "",
+        "Total Cost": "",
+        Comment: "",
+      });
+    } else {
+      alert("Record not updated!");
+    }
   };
   const columnNames = [
     "Year",
@@ -377,9 +401,9 @@ function Encoder() {
   return (
     <div className="App">
       <h1>DATA ENCODER</h1>
-      <div >
+      <div>
         <h2>Add Record</h2>
-        <form onSubmit={createNote} >
+        <form onSubmit={createNote}>
           <input
             onChange={updateCreateForm}
             name="Year"
@@ -512,7 +536,7 @@ function Encoder() {
             placeholder="Cost of Water (Sweat) Purchase_Treatment_Water Waste_Disposal"
             value={
               createForm[
-              "Cost of Water (Sweat) Purchase_Treatment_Water Waste_Disposal"
+                "Cost of Water (Sweat) Purchase_Treatment_Water Waste_Disposal"
               ]
             }
           />
@@ -628,15 +652,15 @@ function Encoder() {
         </form>
       </div>
 
-      <div className="records" >
-
+      <div className="records">
         <h2>Records</h2>
 
         <div className="excel-header">
-
           {columnNames.map((columnName, index) => (
             <div key={index} className="box" style={{ color: "white" }}>
-              <h4 style={{ margin: "0", whiteSpace: "nowrap" }} >{columnName}</h4>
+              <h4 style={{ margin: "0", whiteSpace: "nowrap" }}>
+                {columnName}
+              </h4>
             </div>
           ))}
         </div>
@@ -649,34 +673,49 @@ function Encoder() {
               .reduce((acc, [key, value]) => ({ ...acc, [key]: value }), {});
 
             return (
-
               <div className="record-2" key={record._id}>
-
                 {Object.entries(filteredRecord).map(
                   ([heading, value], index) => (
-                    <div
-                      key={index}
-                      className="box">
-                      <h4 style={{ color: "white", margin: "0", whiteSpace: "nowrap" }}>{heading}</h4>
-                      <p >{value}</p>
+                    <div key={index} className="box">
+                      <h4
+                        style={{
+                          color: "white",
+                          margin: "0",
+                          whiteSpace: "nowrap",
+                        }}
+                      >
+                        {heading}
+                      </h4>
+                      <p>{value}</p>
                     </div>
                   )
                 )}
-                <button className="edit-button" onClick={() => updateNote(record)}> <svg class="edit-svgIcon" viewBox="0 0 512 512">
-                  <path d="M410.3 231l11.3-11.3-33.9-33.9-62.1-62.1L291.7 89.8l-11.3 11.3-22.6 22.6L58.6 322.9c-10.4 10.4-18 23.3-22.2 37.4L1 480.7c-2.5 8.4-.2 17.5 6.1 23.7s15.3 8.5 23.7 6.1l120.3-35.4c14.1-4.2 27-11.8 37.4-22.2L387.7 253.7 410.3 231zM160 399.4l-9.1 22.7c-4 3.1-8.5 5.4-13.3 6.9L59.4 452l23-78.1c1.4-4.9 3.8-9.4 6.9-13.3l22.7-9.1v32c0 8.8 7.2 16 16 16h32zM362.7 18.7L348.3 33.2 325.7 55.8 314.3 67.1l33.9 33.9 62.1 62.1 33.9 33.9 11.3-11.3 22.6-22.6 14.5-14.5c25-25 25-65.5 0-90.5L453.3 18.7c-25-25-65.5-25-90.5 0zm-47.4 168l-144 144c-6.2 6.2-16.4 6.2-22.6 0s-6.2-16.4 0-22.6l144-144c6.2-6.2 16.4-6.2 22.6 0s6.2 16.4 0 22.6z"></path>
-                </svg></button>
-                <button className="delete-btn" onClick={() => deleteNote(record._id)}><svg class="delete-svgIcon" viewBox="0 0 448 512">
-                  <path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"></path>
-                </svg>  </button>
+                <button
+                  className="edit-button"
+                  onClick={() => updateNote(record)}
+                >
+                  {" "}
+                  <svg class="edit-svgIcon" viewBox="0 0 512 512">
+                    <path d="M410.3 231l11.3-11.3-33.9-33.9-62.1-62.1L291.7 89.8l-11.3 11.3-22.6 22.6L58.6 322.9c-10.4 10.4-18 23.3-22.2 37.4L1 480.7c-2.5 8.4-.2 17.5 6.1 23.7s15.3 8.5 23.7 6.1l120.3-35.4c14.1-4.2 27-11.8 37.4-22.2L387.7 253.7 410.3 231zM160 399.4l-9.1 22.7c-4 3.1-8.5 5.4-13.3 6.9L59.4 452l23-78.1c1.4-4.9 3.8-9.4 6.9-13.3l22.7-9.1v32c0 8.8 7.2 16 16 16h32zM362.7 18.7L348.3 33.2 325.7 55.8 314.3 67.1l33.9 33.9 62.1 62.1 33.9 33.9 11.3-11.3 22.6-22.6 14.5-14.5c25-25 25-65.5 0-90.5L453.3 18.7c-25-25-65.5-25-90.5 0zm-47.4 168l-144 144c-6.2 6.2-16.4 6.2-22.6 0s-6.2-16.4 0-22.6l144-144c6.2-6.2 16.4-6.2 22.6 0s6.2 16.4 0 22.6z"></path>
+                  </svg>
+                </button>
+                <button
+                  className="delete-btn"
+                  onClick={() => deleteNote(record._id)}
+                >
+                  <svg class="delete-svgIcon" viewBox="0 0 448 512">
+                    <path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"></path>
+                  </svg>{" "}
+                </button>
               </div>
             );
           })}
       </div>
 
       {updateForm._id && (
-        <div >
+        <div>
           <h2>Update Record</h2>
-          <form onSubmit={update} >
+          <form onSubmit={update}>
             <input
               onChange={handleUpdateFieldChange}
               name="Year"
@@ -767,7 +806,7 @@ function Encoder() {
               placeholder="Hazardous Waste (Maintenace Oil)(With Density 0_93)"
               value={
                 updateForm[
-                "Hazardous Waste (Maintenace Oil)(With Density 0_93)"
+                  "Hazardous Waste (Maintenace Oil)(With Density 0_93)"
                 ]
               }
             />
@@ -835,13 +874,16 @@ function Encoder() {
               value={updateForm["Total Cost"]}
             />
 
-            <button className="btn" type="submit">Update Record</button>
+            <button className="btn" type="submit">
+              Update Record
+            </button>
           </form>
         </div>
       )}
-      <div style={{display:"flex"}}>
-      <button className="nav-btn" onClick={null}>Send Data</button>
-      
+      <div style={{ display: "flex" }}>
+        <button className="nav-btn" onClick={send_data}>
+          Send Data
+        </button>
       </div>
     </div>
   );
